refactor: API, 컴포넌트 관련 코드 리뷰 피드백 반영 #157
