+++ src/App.tsx
@@ -14,7 +14,6 @@ import GlobalStyles from '@/styles/GlobalStyles';
 import Layout from '@/layout/Layout';
 import Login from '@/pages/Login';
 import UnderConstruction from '@/pages/UnderConstruction';
-import ScrollToTop from '@/components/ScrollToTop';
 import NotFoundPage from './pages/NotFoundPage';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
@@ -30,7 +29,6 @@ const PrivateRoute = () => {
 const App = () => (
 	<BrowserRouter>
 		<GlobalStyles />
-		<ScrollToTop />
 		<QueryClientProvider client={queryClient}>
 			<Routes>
 				<Route path="/" element={<Layout />}>
