+++ src/api/schedule/deletePersonalSchedule.ts
@@ -1,6 +1,6 @@
 import { db } from '@/firebaseConfig';
 import { collection, query, where, getDocs, updateDoc, deleteDoc } from 'firebase/firestore';
-import getUserId from '@/api/common/getUserId';
+import { getUserId } from '@/utils/userDataUtils';
 
 type TWorkingTimes = 'open' | 'middle' | 'close';
 
@@ -18,7 +18,7 @@ const deletePersonalSchedule = async (workDate: string, workingTimes: TWorkingTi
 
 		const querySnapshot = await getDocs(q);
 
-		querySnapshot.forEach(async (doc) => {
+		for (const doc of querySnapshot.docs) {
 			const data = doc.data();
 			const workingTimesArray = data.workingTimes || [];
 			const memosArray = data.memos || [];
@@ -39,7 +39,7 @@ const deletePersonalSchedule = async (workDate: string, workingTimes: TWorkingTi
 					});
 				}
 			}
-		});
+		}
 
 		return true;
 	} catch (error) {
