+++ src/api/schedule/getOfficialSchedule.ts
@@ -1,6 +1,6 @@
 import { db } from '@/firebaseConfig';
 import { collection, query, where, getDocs } from 'firebase/firestore';
-import getUserId from '@/api/common/getUserId';
+import { getUserId } from '@/utils/userDataUtils';
 
 // 특정 달의 공식 근무 일정 조회 API
 const getOfficialSchedule = async (year: number, month: number) => {
