+++ src/api/user/getUserInfo.ts
@@ -1,6 +1,6 @@
 import { db } from '@/firebaseConfig';
 import { collection, query, where, getDocs } from 'firebase/firestore';
-import getUserId from '@/api/common/getUserId';
+import { getUserId } from '@/utils/userDataUtils';
 
 // 회원 정보 조회 API
 const getUserInfo = async (): Promise<{
