+++ src/api/work/getCorrection.ts
@@ -1,6 +1,6 @@
 import { db } from '@/firebaseConfig';
 import { collection, query, where, getDocs } from 'firebase/firestore';
-import getUserId from '@/api/common/getUserId';
+import { getUserId } from '@/utils/userDataUtils';
 
 // 근무 정정신청 내역 조회 API
 const getCorrection = async () => {
