+++ src/api/work/getPersonalWage.ts
@@ -1,6 +1,6 @@
 import { db } from '@/firebaseConfig';
 import { collection, query, where, getDocs } from 'firebase/firestore';
-import getUserId from '@/api/common/getUserId';
+import { getUserId } from '@/utils/userDataUtils';
 
 // 개인 스케줄에 따른 특정 달의 예상 급여액 조회 API
 const getPersonalWage = async (year: number, month: number) => {
