+++ src/utils/scrollUtils.ts
@@ -1,0 +1,4 @@
+// 스크롤을 최상단으로 초기화시키는 로직
+export const ScrollToTop = (): void => {
+	window.scrollTo(0, 0);
+};
