+++ src/utils/userDataUtils.ts
@@ -1,5 +1,5 @@
 // 현재 로그인한 사용자의 UID를 반환하는 로직
-const getUserId = (): string => {
+export const getUserId = (): string => {
 	const sessionKey = `firebase:authUser:${import.meta.env.VITE_FIREBASE_API_KEY as string}:[DEFAULT]`;
 	const userSession = sessionStorage.getItem(sessionKey);
 
@@ -10,5 +10,3 @@ const getUserId = (): string => {
 		return '';
 	}
 };
-
-export default getUserId;
